refactor: 💡 cerate hook get registrations && update

+++ src/hooks/useRegistration.ts
@@ -1,16 +1,45 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import { StatusEnum } from "~/enums/StatusEnum";
 import { api } from "~/services";
 import { RegistrationData } from "~/types/RegistrationCard";
 import useNotify from "./useNotify";
 
-const useRegistrationActions = (data: RegistrationData, updateData: any) => {
+const useRegistrationActions = (data?: RegistrationData, updateData?: () => void) => {
   const [modalVisible, setModalVisible] = useState(false);
   const [modalAction, setModalAction] = useState('');
+  const [registrations, setRegistrations] = useState<RegistrationData[]>([]);
+  const [isLoading, setIsLoading] = useState(true);
   const { notification, notifySuccess } = useNotify();
 
+  useEffect(() => {
+    const getData = async () => {
+      try {
+        const response = await api.getRegistrations();
+        setRegistrations(response);
+        setIsLoading(false);
+      } catch (error) {
+        console.error('Error fetching registrations:', error);
+        setIsLoading(false);
+      }
+    };
+
+    getData();
+  }, []);
+
+  const updateRegistrations = async () => {
+    setIsLoading(true);
+    try {
+      const response = await api.getRegistrations();
+      setRegistrations(response);
+    } catch (error) {
+      console.error('Error updating registrations:', error);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
   const handleStatusChange = async (newStatus: StatusEnum) => {
-    if (data.status !== newStatus) {
+    if (data && data.status !== newStatus) {
       try {
         await api.updateRegistration(data.id, {
           ...data,
@@ -18,7 +47,7 @@ const useRegistrationActions = (data: RegistrationData, updateData: any) => {
         });
 
         notifySuccess(`Status alterado para ${newStatus}`);
-        updateData();
+        if (updateData) updateData();
         
       } catch (error) {
         console.error(error);
@@ -27,13 +56,15 @@ const useRegistrationActions = (data: RegistrationData, updateData: any) => {
   };
 
   const handleStatusDelete = async () => {
-    try {
-      await api.deleteRegistration(data.id);
-      notifySuccess('Registro excluído com sucesso');
-      updateData();
+    if (data) {
+      try {
+        await api.deleteRegistration(data.id);
+        notifySuccess('Registro excluído com sucesso');
+        if (updateData) updateData();
 
-    } catch (error) {
-      console.error(error);
+      } catch (error) {
+        console.error(error);
+      }
     }
   };
 
@@ -63,9 +94,12 @@ const useRegistrationActions = (data: RegistrationData, updateData: any) => {
   return {
     notification,
     modalVisible,
+    registrations, 
+    isLoading, 
     handleOpenModal,
     handleCloseModal,
-    handleConfirmAction
+    handleConfirmAction,
+    updateRegistrations
   };
 };
